backend/requirements.txt
fastapi
uvicorn
vaderSentiment
networkx
python-dotenv
mistralai
